DetectTOC: only go for lines containing '...' words
DetectTOC will work on all lines, shaving off numbers from the last of
the words in a given line, so long as the word is not all full-stops.
This implies that a TOC line is one that contains strings containing
only full-stops, and so, DetectTOC should only work on such lines.

This change will remove unwanted behaviour where DetectTOC removes
trailing numbers that we actually want to keep in lines, eg:

  Case Number : ABC 12/1234